refactor(logic): split monolithic logic.ts into domain modules
Extract env.ts (constants, resolveApiBase, sanitizeApiKey), auth.ts
(resolveApiKey), models.ts (MODELS, fetchRemoteModels, resolveModels),
errors.ts (ClinePassErrorType, classifyClinePassError),
error-handler.ts (handleClinePassError handler), utils.ts (type
guards), and workos.ts (resolveClineAuthCredentials, isWorkosToken).

Each module has its own test file. The message_end handler moved from
inline closure in index.ts to a standalone function in
error-handler.ts, making it testable in isolation.
