+++ tests/unit/index.test.ts
@@ -1,31 +1,6 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
-import {
-  DEFAULT_API_BASE,
-  ENV_API_KEY,
-  PROVIDER_NAME,
-  MODELS,
-  CLINEPASS_ERROR_MESSAGES,
-} from "../../src/logic.js";
-
-/** Shared handler type for the message_end event in tests. */
-type MessageEndHandler = (
-  event: { message: { stopReason?: string; errorMessage?: string; provider?: string } },
-  ctx: {
-    hasUI: boolean;
-    ui: { notify: (msg: string, type: string) => void };
-    model?: { provider?: string };
-  },
-) => void;
-
-/** Minimal fake pi with both registerProvider and on() for handler tests. */
-function makeFakePi(handlerRef: { h?: MessageEndHandler }) {
-  return {
-    registerProvider(_name: string, _config: Record<string, unknown>) {},
-    on(event: string, h: MessageEndHandler) {
-      if (event === "message_end") handlerRef.h = h;
-    },
-  };
-}
+import { DEFAULT_API_BASE, ENV_API_KEY, PROVIDER_NAME } from "../../src/env.js";
+import { MODELS } from "../../src/models.js";
 
 /**
  * Verifies the provider registration shape passed to pi.registerProvider.
@@ -117,10 +92,10 @@ describe("provider registration", () => {
 });
 
 /**
- * Verifies the message_end error handler surfaces friendly messages for
- * ClinePass-specific errors (403, 401, 429).
+ * Verifies that the extension registers a message_end event listener.
+ * The handler logic itself is tested in error-handler.test.ts.
  */
-describe("message_end error handler", () => {
+describe("message_end event registration", () => {
   beforeEach(() => {
     vi.stubGlobal("fetch", vi.fn().mockResolvedValue(new Response("Not Found", { status: 404 })));
   });
@@ -143,200 +118,5 @@ describe("message_end error handler", () => {
     await mod.default(fakePi as never);
 
     expect(registeredEvents).toContain("message_end");
-  });
-
-  it("surfaces friendly message for 403 errors from clinepass", async () => {
-    const ref: { h?: MessageEndHandler } = {};
-    const notifyCalls: { msg: string; type: string }[] = [];
-
-    const mod = await import("../../src/index.js");
-    await mod.default(makeFakePi(ref) as never);
-
-    expect(ref.h).toBeDefined();
-    ref.h!(
-      {
-        message: {
-          stopReason: "error",
-          errorMessage: "Request failed with status 403",
-          provider: PROVIDER_NAME,
-        },
-      },
-      {
-        hasUI: true,
-        ui: { notify: (msg: string, type: string) => notifyCalls.push({ msg, type }) },
-        model: { provider: PROVIDER_NAME },
-      },
-    );
-
-    expect(notifyCalls).toHaveLength(1);
-    expect(notifyCalls[0].msg).toBe(CLINEPASS_ERROR_MESSAGES.not_subscribed);
-    expect(notifyCalls[0].type).toBe("error");
-  });
-
-  it("surfaces friendly message for 401 errors", async () => {
-    const ref: { h?: MessageEndHandler } = {};
-    const notifyCalls: { msg: string; type: string }[] = [];
-
-    const mod = await import("../../src/index.js");
-    await mod.default(makeFakePi(ref) as never);
-
-    ref.h!(
-      {
-        message: { stopReason: "error", errorMessage: "401 Unauthorized", provider: PROVIDER_NAME },
-      },
-      {
-        hasUI: true,
-        ui: { notify: (msg: string, type: string) => notifyCalls.push({ msg, type }) },
-        model: { provider: PROVIDER_NAME },
-      },
-    );
-
-    expect(notifyCalls).toHaveLength(1);
-    expect(notifyCalls[0].msg).toBe(CLINEPASS_ERROR_MESSAGES.auth_expired);
-  });
-
-  it("surfaces friendly message for 429 rate limit errors", async () => {
-    const ref: { h?: MessageEndHandler } = {};
-    const notifyCalls: { msg: string; type: string }[] = [];
-
-    const mod = await import("../../src/index.js");
-    await mod.default(makeFakePi(ref) as never);
-
-    ref.h!(
-      {
-        message: {
-          stopReason: "error",
-          errorMessage: "429 Too Many Requests",
-          provider: PROVIDER_NAME,
-        },
-      },
-      {
-        hasUI: true,
-        ui: { notify: (msg: string, type: string) => notifyCalls.push({ msg, type }) },
-        model: { provider: PROVIDER_NAME },
-      },
-    );
-
-    expect(notifyCalls).toHaveLength(1);
-    expect(notifyCalls[0].msg).toBe(CLINEPASS_ERROR_MESSAGES.rate_limited);
-  });
-
-  it("uses ctx.model.provider when message has no provider field", async () => {
-    const ref: { h?: MessageEndHandler } = {};
-    const notifyCalls: { msg: string; type: string }[] = [];
-
-    const mod = await import("../../src/index.js");
-    await mod.default(makeFakePi(ref) as never);
-
-    ref.h!(
-      {
-        message: { stopReason: "error", errorMessage: "403 Forbidden" },
-      },
-      {
-        hasUI: true,
-        ui: { notify: (msg: string, type: string) => notifyCalls.push({ msg, type }) },
-        model: { provider: PROVIDER_NAME },
-      },
-    );
-
-    expect(notifyCalls).toHaveLength(1);
-    expect(notifyCalls[0].msg).toBe(CLINEPASS_ERROR_MESSAGES.not_subscribed);
-  });
-
-  it("surfaces friendly message for unknown error types", async () => {
-    const ref: { h?: MessageEndHandler } = {};
-    const notifyCalls: { msg: string; type: string }[] = [];
-
-    const mod = await import("../../src/index.js");
-    await mod.default(makeFakePi(ref) as never);
-
-    ref.h!(
-      {
-        message: {
-          stopReason: "error",
-          errorMessage: "Internal server error",
-          provider: PROVIDER_NAME,
-        },
-      },
-      {
-        hasUI: true,
-        ui: { notify: (msg: string, type: string) => notifyCalls.push({ msg, type }) },
-        model: { provider: PROVIDER_NAME },
-      },
-    );
-
-    expect(notifyCalls).toHaveLength(1);
-    expect(notifyCalls[0].msg).toBe(CLINEPASS_ERROR_MESSAGES.unknown);
-    expect(notifyCalls[0].type).toBe("error");
-  });
-
-  it("ignores errors from other providers", async () => {
-    const ref: { h?: MessageEndHandler } = {};
-    const notifyCalls: { msg: string; type: string }[] = [];
-
-    const mod = await import("../../src/index.js");
-    await mod.default(makeFakePi(ref) as never);
-
-    ref.h!(
-      {
-        message: { stopReason: "error", errorMessage: "403 Forbidden", provider: "openai" },
-      },
-      {
-        hasUI: true,
-        ui: { notify: (msg: string, type: string) => notifyCalls.push({ msg, type }) },
-        model: { provider: "openai" },
-      },
-    );
-
-    expect(notifyCalls).toHaveLength(0);
-  });
-
-  it("ignores non-error messages", async () => {
-    const ref: { h?: MessageEndHandler } = {};
-    const notifyCalls: { msg: string; type: string }[] = [];
-
-    const mod = await import("../../src/index.js");
-    await mod.default(makeFakePi(ref) as never);
-
-    ref.h!(
-      {
-        message: { stopReason: "stop", provider: PROVIDER_NAME },
-      },
-      {
-        hasUI: true,
-        ui: { notify: (msg: string, type: string) => notifyCalls.push({ msg, type }) },
-        model: { provider: PROVIDER_NAME },
-      },
-    );
-
-    expect(notifyCalls).toHaveLength(0);
-  });
-
-  it("falls back to console.error when no UI", async () => {
-    const ref: { h?: MessageEndHandler } = {};
-    const errorSpy = vi.spyOn(console, "error").mockImplementation(() => {});
-
-    const mod = await import("../../src/index.js");
-    await mod.default(makeFakePi(ref) as never);
-
-    ref.h!(
-      {
-        message: {
-          stopReason: "error",
-          errorMessage: "403 Forbidden",
-          provider: PROVIDER_NAME,
-        },
-      },
-      {
-        hasUI: false,
-        ui: { notify: () => {} },
-        model: { provider: PROVIDER_NAME },
-      },
-    );
-
-    expect(errorSpy).toHaveBeenCalledTimes(1);
-    expect(errorSpy.mock.calls[0][0]).toContain("[clinepass]");
-    expect(errorSpy.mock.calls[0][0]).toContain(CLINEPASS_ERROR_MESSAGES.not_subscribed);
-    errorSpy.mockRestore();
   });
 });
